+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - aligned unsafe-io file name to corresponding resource file
+- logging OS errors on executing command
 
 ## [0.2.0] - 2018-06-04
 ### Added